slider added. small bug on the values it can take on

+++ ConfigSetup.h
@@ -4,7 +4,7 @@
 #include "emp/config/ArgManager.hpp"
 
 EMP_BUILD_CONFIG(MyConfigType,
-    VALUE(LUMINOSITY, int, 0, "What value should the luminosity be?")
+    VALUE(LUMINOSITY, float, 0, "What value should the luminosity be?")
 )
 
 #endif
